requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
